Update thorlabs_cube_motor_kinesis.py
Add steppermotor support (ZST213B) to thorlabs_cube_motor_kinesis.
The DLL methods have a different prefix compared to DC motors, which requires a lot of if ... else when addressing the motors.
I could not get the GetDeviceUnitFromRealValue method  to work, so instead I created methods directly in the python class to do the conversions.